% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimator.R
\name{bayes_estimator}
\alias{bayes_estimator}
\title{Add together two numbers}
\usage{
bayes_estimator(formula, data, family, summary_probs = c(0.1, 0.5, 0.9))
}
\arguments{
\item{formula}{A model formula.}

\item{data}{A data.frame containing the data.}

\item{family}{The model family. Currently implemented are 'binomial' and 'poisson'.}

\item{summary_probs}{Vector containing the quantiles for summarizing the joint posterior distribution.}
}
\value{
A sumary of the joint posterior distribution.
}
\description{
Add together two numbers
}
