% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Funktion for comparing models}
\usage{
compare_models(newdata, models)
}
\arguments{
\item{newdata}{Newdata.}

\item{models}{Models.}
}
\description{
Funktion for comparing models
}
