% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance_summary.R
\name{disturbance_summary}
\alias{disturbance_summary}
\title{Function for summarizing disturbances from a TimeSync sample}
\usage{
disturbance_summary(dat, by.year = TRUE, by.agent = FALSE,
  grouping.vars = NULL, sub.agents = FALSE, change.agents = c("Harvest",
  "Wind", "Decline", "Hydrology", "Debris", "Other"), agent.regroup = NULL,
  interpreter = NULL, remove.1985.decline = TRUE)
}
\arguments{
\item{dat}{Data exported from TimeSync.}

\item{by.year}{TBD...}

\item{by.agent}{Logical. Whether disturbances should be aggregated by agent.}

\item{grouping.vars}{Data.frame. Additional grouping variables stored in a data frame, where each row corresponds to one plot. The data frame must have a column 'plotid'.}

\item{sub.agents}{TBD...}

\item{change.agents}{Change agents to be considered. Standard is 'Harvest', 'Wind', and 'Decline'.}

\item{agent.regroup}{TBD...}

\item{interpreter}{TBD...}

\item{remove.1985.decline}{TBD...}
}
\value{
A sumary of the disturbances per year (and agents/grouping variables).
}
\description{
Function for summarizing disturbances from a TimeSync sample
}
