\name{disturbanceBayes-package}
\alias{disturbanceBayes-package}
\alias{disturbanceBayes}
\docType{package}
\title{
\packageTitle{disturbanceBayes}
}
\description{
\packageDescription{disturbanceBayes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{disturbanceBayes}
\packageIndices{disturbanceBayes}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{disturbanceBayes}

Maintainer: \packageMaintainer{disturbanceBayes}
}
\references{
 Literature or other references for background information
}
 Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
 Optional links to other man pages, e.g.
 \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{

}
