% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_estimator.R
\name{bayes_estimator}
\alias{bayes_estimator}
\title{Function for estimating disturbance rates from TimeSync samples using Bayesian hierarchical modeling}
\usage{
bayes_estimator(x, p, index_cols, model)
}
\arguments{
\item{x}{A data frame returned from \code{disturbance_summary()}.}

\item{p}{A vector indicating the quantiles used for summarizing the posterior.}

\item{index_cols}{A vector indexing the columns storing the hierarchical levels (e.g., years). Either index value or column name.}

\item{model}{The model family. Either 'binomial' or 'poisson'.}
}
\value{
A sumary of the joint posterior distribution for each hierarchical level.
}
\description{
Function for estimating disturbance rates from TimeSync samples using Bayesian hierarchical modeling
}
