% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Add together two numbers}
\usage{
compare_models(newdata, models)
}
\arguments{
\item{formula}{A model formula.}

\item{data}{A data.frame containing the data.}

\item{family}{The model family. Currently implemented are 'binomial' and 'poisson'.}

\item{summary_probs}{Vector containing the quantiles for summarizing the joint posterior distribution.}
}
\value{
A sumary of the joint posterior distribution.
}
\description{
Add together two numbers
}
